requirements.txt
aiohttp==3.6.2
akinator.py==4.0.1
alembic==1.4.3
async-timeout==3.0.1
asyncssh==2.4.2
attrs==20.1.0
beautifulsoup4==4.9.3
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
cssselect2==0.3.0
defusedxml==0.6.0
discord.py==1.4.1
flatlib==0.2.1
geographiclib==1.50
geopy==2.0.0
greenlet==1.0.0
idna==2.10
Mako==1.1.3
mal-api==0.3.1
MarkupSafe==1.1.1
multidict==4.7.6
numpy==1.19.2
Pillow==7.2.0
psycopg2-binary==2.8.6
pycparser==2.20
pyswisseph==2.0.0.post2
python-chess==0.31.4
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.4.0b1
timezonefinder==4.4.1
tinycss2==1.0.2
urllib3==1.25.10
vcrpy==4.1.1
vcrpy-unittest==0.1.7
webencodings==0.5.1
wrapt==1.12.1
yarl==1.5.1
